Move parquet formatter makeUpdatedLocation to base class
This code was being used a lot and the method is generally
useful and will simplify datastore.getUri
